backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-socketio>=5.11.0
pydantic>=2.6.0
python-multipart>=0.0.9
opencv-python-headless>=4.8.0
numpy>=1.24.0
ultralytics>=8.3.0
tensorflow>=2.15.0
mediapipe>=0.10.0
requests>=2.31.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
slowapi>=0.1.0
